Fixed Formant Filters
The problem arose because the radius of the filter was set to 0, this was normalizing all of the sounds, making them all sound the exact same. The radius is now .9 and all of the filters sound different (in a way that I assume is at least similar to how they were meant to sound).
One thing that I have noticed is that pressing multiple hiss keys at the same time seems to create a strange bit of extra noise that sounds a bit bad, buzz seems to work fine though.